Percentile Bootstrap CIs for Standardized Solution

It receives a lavaan::lavaan object fitted with bootstrapping standard errors requested and forms the percentile confidence intervals for the standardized solution.

standardizedSolution_boot_ci(object, level = 0.95, type = "std.all", ...)

Arguments

object

A lavaan object, fitted with 'se = "boot"'.

level

The level of confidence of the confidence intervals. Default is .95.

type

The type of standard estimates. The same argument of lavaan::standardizedSolution(), and support all values supported by lavaan::standardizedSolution(). Default is "std.all".

...

Other arguments to be passed to lavaan::standardizedSolution().

Value

The output of lavaan::standardizedSolution(), with percentile bootstrap confidence intervals appended to the right.

Author

Shu Fai Cheung (shufai.cheung@gmail.com)

Examples

library(lavaan)
#> This is lavaan 0.6-11
#> lavaan is FREE software! Please report any bugs.
set.seed(5478374)
n <- 50
x <- runif(n) - .5
m <- .40 * x + rnorm(n, 0, sqrt(1 - .40))
y <- .30 * m + rnorm(n, 0, sqrt(1 - .30))
dat <- data.frame(x = x, y = y, m = m)
model <-
'
m ~ a*x
y ~ b*m
ab := a*b
'

# Should set bootstrap to at least 2000 in real studies
fit <- sem(model, data = dat, fixed.x = FALSE,
           se = "boot",
           bootstrap = 100)
summary(fit)
#> lavaan 0.6-11 ended normally after 1 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#>                                                       
#>   Number of observations                            50
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.020
#>   Degrees of freedom                                 1
#>   P-value (Chi-square)                           0.887
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                            Bootstrap
#>   Number of requested bootstrap draws              100
#>   Number of successful bootstrap draws             100
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   m ~                                                 
#>     x          (a)    0.569    0.337    1.688    0.091
#>   y ~                                                 
#>     m          (b)    0.219    0.146    1.503    0.133
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .m                 0.460    0.079    5.803    0.000
#>    .y                 0.570    0.104    5.487    0.000
#>     x                 0.078    0.011    7.017    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     ab                0.125    0.137    0.910    0.363
#> 

standardizedSolution_boot_ci(fit)
#> Warning: the condition has length > 1 and only the first element will be used
#>   lhs op rhs label est.std    se      z pvalue ci.lower ci.upper boot.ci.lower
#> 1   m  ~   x     a   0.229 0.136  1.677  0.094   -0.039    0.496        -0.013
#> 2   y  ~   m     b   0.198 0.125  1.586  0.113   -0.047    0.443        -0.020
#> 3   m ~~   m         0.948 0.062 15.205  0.000    0.826    1.070         0.737
#> 4   y ~~   y         0.961 0.050 19.401  0.000    0.864    1.058         0.808
#> 5   x ~~   x         1.000 0.000     NA     NA    1.000    1.000         1.000
#> 6  ab := a*b    ab   0.045 0.046  0.975  0.330   -0.046    0.136        -0.006
#>   boot.ci.upper
#> 1         0.513
#> 2         0.439
#> 3         1.000
#> 4         1.000
#> 5         1.000
#> 6         0.178