Bootstrap CIs for Standardized Solution

It receives a lavaan::lavaan object fitted with bootstrapping standard errors requested and forms the confidence intervals for the standardized solution.

It works by calling lavaan::standardizedSolution() with the bootstrap estimates of free parameters in each bootstrap sample to compute the standardized estimates in each sample.

A more reliable way is to use function like lavaan::bootstrapLavaan(). Nevertheless, this simple function is good enough for some simple scenarios, and does not require repeating the bootstrapping step.

Usage

standardizedSolution_boot_ci(
  object,
  level = 0.95,
  type = "std.all",
  save_boot_est_std = TRUE,
  force_run = FALSE,
  boot_delta_ratio = FALSE,
  ...
)

Arguments

object

A lavaan object, fitted with 'se = "boot"'.

level

The level of confidence of the confidence intervals. Default is .95.

type

The type of standard estimates. The same argument of lavaan::standardizedSolution(), and support all values supported by lavaan::standardizedSolution(). Default is "std.all".

save_boot_est_std

Whether the bootstrap estimates of the standardized solution are saved. If saved, they will be stored in the attribute boot_est_std. Default is TRUE.

force_run

If TRUE, will skip checks and run models without checking the estimates. For internal use. Default is FALSE.

boot_delta_ratio

The ratio of (a) the distance of the bootstrap confidence limit from the point estimate to (b) the distance of the delta-method limit from the point estimate.

...

Other arguments to be passed to lavaan::standardizedSolution().

Value

The output of lavaan::standardizedSolution(), with bootstrap confidence intervals appended to the right.

See also

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448. Originally proposed in an issue at GitHub https://github.com/simsem/semTools/issues/101#issue-1021974657, inspired by a discussion at the Google group for lavaan https://groups.google.com/g/lavaan/c/qQBXSz5cd0o/m/R8YT5HxNAgAJ. boot::boot.ci() is used to form the percentile confidence intervals in this version.

Examples

library(lavaan)
set.seed(5478374)
n <- 50
x <- runif(n) - .5
m <- .40 * x + rnorm(n, 0, sqrt(1 - .40))
y <- .30 * m + rnorm(n, 0, sqrt(1 - .30))
dat <- data.frame(x = x, y = y, m = m)
model <-
'
m ~ a*x
y ~ b*m
ab := a*b
'

# Should set bootstrap to at least 2000 in real studies
fit <- sem(model, data = dat, fixed.x = FALSE,
           se = "boot",
           bootstrap = 100)
summary(fit)
#> lavaan 0.6-12 ended normally after 1 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                            50
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.020
#>   Degrees of freedom                                 1
#>   P-value (Chi-square)                           0.887
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                            Bootstrap
#>   Number of requested bootstrap draws              100
#>   Number of successful bootstrap draws             100
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   m ~                                                 
#>     x          (a)    0.569    0.337    1.688    0.091
#>   y ~                                                 
#>     m          (b)    0.219    0.146    1.503    0.133
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .m                 0.460    0.079    5.803    0.000
#>    .y                 0.570    0.104    5.487    0.000
#>     x                 0.078    0.011    7.017    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     ab                0.125    0.137    0.910    0.363
#> 

standardizedSolution_boot_ci(fit)
#>   lhs op rhs label est.std    se      z pvalue ci.lower ci.upper boot.ci.lower
#> 1   m  ~   x     a   0.229 0.136  1.677  0.094   -0.039    0.496        -0.039
#> 2   y  ~   m     b   0.198 0.125  1.586  0.113   -0.047    0.443        -0.026
#> 3   m ~~   m         0.948 0.062 15.205  0.000    0.826    1.070         0.708
#> 4   y ~~   y         0.961 0.050 19.401  0.000    0.864    1.058         0.796
#> 5   x ~~   x         1.000 0.000     NA     NA    1.000    1.000            NA
#> 6  ab := a*b    ab   0.045 0.046  0.975  0.330   -0.046    0.136        -0.008
#>   boot.ci.upper
#> 1         0.540
#> 2         0.451
#> 3         1.000
#> 4         1.000
#> 5            NA
#> 6         0.188